Cycles: pointcloud geometry type
This add support for rendering of the point cloud object in Blender, as a native
geometry type in Cycles that is more memory and time efficient than instancing
sphere meshes.

Points are currently always rendered as spheres. More shapes may be added later,
but this is the most important one and can be customized with shaders.

For CPU rendering the Embree primitive is used, for GPU there is our own
intersection code. Motion blur is suppored.

The OptiX implementation needs more work. It currently is done as part of the
curve intersection function, however once we use the native OptiX curve
primitive this will not work. Some help with this would be welcome.

I'm planning to add this for 2.93. The new geometry nodes are not meant to
generate point clouds of this type yet (particles, water splashes, sand, etc).
So there is no rush reviewing or committing this.

Differential Revision: https://developer.blender.org/D9887