Decouple transform and deform motion blur steps
Transformation and deformation motion blur were already very
independent, with the exception that they both used the same variable
to hold the number of steps.

They are now separated, deformation blur uses an explicit variable
in ccl::Geometry, transformation blur is implicit in the size of the
ccl::Object::motion array.